requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.6
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
idna==3.2
jwcrypto==1.0
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
wrapt==1.12.1
